A few more static analysis warnings

+++ GPU/Directx9/FramebufferDX9.cpp
@@ -588,7 +588,7 @@ namespace DX9 {
 	}
 
 	FBO *FramebufferManagerDX9::GetTempFBO(u16 w, u16 h, FBOColorDepth depth) {
-		u64 key = ((u64)depth << 32) | (w << 16) | h;
+		u64 key = ((u64)depth << 32) | ((u32)w << 16) | h;
 		auto it = tempFBOs_.find(key);
 		if (it != tempFBOs_.end()) {
 			it->second.last_frame_used = gpuStats.numFlips;
